+++ backend/canisters/group/impl/src/regular_jobs.rs
@@ -1,4 +1,3 @@
-use crate::updates::c2c_migrate_events_to_stable_memory::migrate_events_to_stable_memory_impl;
 use crate::Data;
 use utils::env::Environment;
 use utils::regular_jobs::{RegularJob, RegularJobs};
@@ -7,13 +6,8 @@ use utils::time::MINUTE_IN_MS;
 pub(crate) fn build() -> RegularJobs<Data> {
     let check_cycles_balance = RegularJob::new("Check cycles balance", check_cycles_balance, 5 * MINUTE_IN_MS);
     let retry_deleting_files = RegularJob::new("Retry deleting files", retry_deleting_files, MINUTE_IN_MS);
-    let migrate_chat_events_to_stable_memory = RegularJob::new("Migrate chat events", migrate_chat_events_to_stable_memory, 0);
 
-    RegularJobs::new(vec![
-        check_cycles_balance,
-        retry_deleting_files,
-        migrate_chat_events_to_stable_memory,
-    ])
+    RegularJobs::new(vec![check_cycles_balance, retry_deleting_files])
 }
 
 fn check_cycles_balance(_: &dyn Environment, data: &mut Data) {
@@ -22,8 +16,4 @@ fn check_cycles_balance(_: &dyn Environment, data: &mut Data) {
 
 fn retry_deleting_files(_: &dyn Environment, _: &mut Data) {
     storage_bucket_client::retry_failed();
-}
-
-fn migrate_chat_events_to_stable_memory(_: &dyn Environment, data: &mut Data) {
-    migrate_events_to_stable_memory_impl(data, true);
 }